Only require go.opencensus.io on Linux
