+++ pkg/version/version.go
@@ -16,16 +16,9 @@
 package version
 
 import (
-	"context"
 	"fmt"
 	"runtime"
 	"strings"
-
-	"go.opencensus.io/stats"
-	"go.opencensus.io/stats/view"
-	"go.opencensus.io/tag"
-
-	"istio.io/istio/pkg/log"
 )
 
 // The following fields are populated at build time using -ldflags -X.
@@ -38,15 +31,6 @@ var (
 	buildDockerHub   = "unknown"
 	buildStatus      = "unknown"
 	buildTag         = "unknown"
-)
-
-var (
-	gitTagKey       tag.Key
-	componentTagKey tag.Key
-	istioBuildTag   = stats.Int64(
-		"istio/build",
-		"Istio component build info",
-		stats.UnitDimensionless)
 )
 
 // BuildInfo describes version information about the binary build.
@@ -139,21 +123,6 @@ func (b BuildInfo) LongForm() string {
 	return fmt.Sprintf("%#v", b)
 }
 
-// RecordComponentBuildTag sets the value for a metric that will be used to track component build tags for
-// tracking rollouts, etc.
-func (b BuildInfo) RecordComponentBuildTag(component string) {
-	b.recordBuildTag(component, tag.New)
-}
-
-func (b BuildInfo) recordBuildTag(component string, newTagCtxFn func(context.Context, ...tag.Mutator) (context.Context, error)) {
-	ctx := context.Background()
-	var err error
-	if ctx, err = newTagCtxFn(ctx, tag.Insert(gitTagKey, b.GitTag), tag.Insert(componentTagKey, component)); err != nil {
-		log.Errorf("could not establish build and component tag keys in context: %v", err)
-	}
-	stats.Record(ctx, istioBuildTag.M(1))
-}
-
 func init() {
 	Info = BuildInfo{
 		Version:       buildVersion,
@@ -165,25 +134,5 @@ func init() {
 		BuildStatus:   buildStatus,
 		GitTag:        buildTag,
 	}
-
-	registerStats(tag.NewKey)
 }
 
-func registerStats(newTagKeyFn func(string) (tag.Key, error)) {
-	var err error
-	if gitTagKey, err = newTagKeyFn("tag"); err != nil {
-		panic(err)
-	}
-	if componentTagKey, err = newTagKeyFn("component"); err != nil {
-		panic(err)
-	}
-	gitTagView := &view.View{
-		Measure:     istioBuildTag,
-		TagKeys:     []tag.Key{componentTagKey, gitTagKey},
-		Aggregation: view.LastValue(),
-	}
-
-	if err = view.Register(gitTagView); err != nil {
-		panic(err)
-	}
-}
